# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  httpadapter.createTriajeRequest:
    properties:
      apellidoMaterno:
        type: string
      apellidoPaterno:
        type: string
      direccion:
        type: string
      edadGestacional:
        type: integer
      email:
        type: string
      esGestante:
        type: boolean
      escalaDolor:
        type: integer
      escalaGlasgow:
        type: integer
      fechaNacimiento:
        type: string
      fiO2:
        type: number
      foto:
        type: string
      fpp:
        type: string
      frecCardiaca:
        type: integer
      frecRespiratoria:
        type: integer
      fur:
        type: string
      gestante:
        type: integer
      idCausaExternaMorbilidad:
        type: integer
      idComunidadDomicilio:
        type: integer
      idDepartamentoDomicilio:
        type: integer
      idDistritoDomicilio:
        type: integer
      idDocIdentidad:
        type: integer
      idEmpleado:
        type: integer
      idEsAccidenteTransito:
        type: integer
      idEstadoCivil:
        type: integer
      idEstadollego:
        type: integer
      idFuenteFinanciamiento:
        type: integer
      idProvinciaDomicilio:
        type: integer
      idServicio:
        type: integer
      idSexo:
        type: integer
      idTipoPrioridad:
        type: integer
      idTriaje:
        type: integer
      imc:
        type: number
      motivo:
        type: string
      movimientosFetales:
        type: integer
      nroControlesPrenatales:
        type: integer
      nroDocumento:
        type: string
      peso:
        type: number
      presionArterial:
        type: string
      primerNombre:
        type: string
      saturacion:
        type: integer
      segundoNombre:
        type: string
      talla:
        type: number
      telefono:
        type: string
      temperatura:
        type: number
      tercerNombre:
        type: string
      tiempoEvolucionCantidad:
        type: integer
      tiempoEvolucionCantidadUnidad:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API del sistema hospitalario Galenos Pro. Arquitectura hexagonal,
    persiste en SQL Server 2022.
  title: Galenos Pro Appointments API
  version: "1.0"
paths:
...
  /triaje/{id}:
    get:
      consumes:
      - application/json
      description: Devuelve los datos del paciente (solo lectura) y del triaje de
        emergencia del id indicado (SP usp_go_Triaje_EmergeciaPorId)
      parameters:
      - description: Id del triaje
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Datos del paciente y del triaje
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Error de validación
          schema:
            type: object
        "404":
          description: Triaje no encontrado
          schema:
            type: object
      summary: Datos de un triaje por id
    put:
      consumes:
      - application/json
      description: Modifica los datos clínicos de un triaje de emergencia invocando
        el SP usp_go_ModificarTriaje
      parameters:
      - description: Id del triaje
        in: path
        name: id
        required: true
        type: integer
      - description: Datos clínicos editables del triaje
        in: body
        name: triaje
        required: true
        schema:
          $ref: '#/definitions/httpadapter.createTriajeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Mensaje de resultado del SP
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Error de validación
          schema:
            type: object
        "500":
          description: Error interno
          schema:
            type: object
      summary: Modifica un triaje
...
  /triaje/referencias:
    get:
      consumes:
      - application/json
      description: Lista las referencias hechas desde consulta externa invocando el
        SP usp_go_ListarBandejaReferencia
      parameters:
      - description: Fecha inicial (YYYY-MM-DD)
        in: query
        name: fini
        required: true
        type: string
      - description: Fecha final (YYYY-MM-DD)
        in: query
        name: ffin
        required: true
        type: string
      - description: Texto de búsqueda por paciente o número de cuenta
        in: query
        name: filtro
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Referencias de consulta externa
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Error de validación
          schema:
            type: object
        "500":
          description: Error interno
          schema:
            type: object
      summary: Bandeja de referencias de consulta externa
  /triaje/referencias/{idAtencion}:
    get:
      consumes:
      - application/json
      description: Devuelve los datos del destino de una referencia (establecimiento,
        UPS y especialidad) invocando el SP usp_go_ListarDatosReferencia
      parameters:
      - description: Id de la atención
        in: path
        name: idAtencion
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Datos de la referencia
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Error de validación
          schema:
            type: object
        "500":
          description: Error interno
          schema:
            type: object
      summary: Datos de una referencia
  /triaje/reporte-por-empleado:
    get:
      consumes:
      - application/json
      description: Cantidad de triajes de un empleado por servicio (tópico) entre
        fechas con hora y minuto (SP usp_go_ReporteTriaje)
      parameters:
      - description: Id del empleado
        in: query
        name: IdEmpleado
        required: true
        type: integer
      - description: Fecha inicial (YYYY-MM-DD o YYYY-MM-DD HH:MM)
        in: query
        name: fechaini
        required: true
        type: string
      - description: Fecha final (YYYY-MM-DD o YYYY-MM-DD HH:MM)
        in: query
        name: fechafin
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cantidad de triajes por tópico
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Error de validación
          schema:
            type: object
      summary: Reporte de triajes por empleado
swagger: "2.0"